数学方法：随机数：Math.random()

+++ src/cn/gtmap/number_and_string/math_method/Test02.java
@@ -6,7 +6,7 @@ public class Test02 {
 		//得到一个0-1之间的随机浮点数（取不到1）
 		System.out.println(Math.random());
 		//得到一个0-10之间的随机浮点数（取不到10）
-		System.out.println((Math.random()*10));
+		System.out.println((int)(Math.random()*10));
 		
 	}
 }
